Let's try fully qualified queue names

+++ integration-test/job-store-service/src/test/java/dk/dbc/dataio/jobstore/EmptyJobsIT.java
@@ -96,7 +96,7 @@ public class EmptyJobsIT extends AbstractJobStoreServiceContainerTest {
         final MockedJmsTextMessage jmsMessage = jmsMessages.get(0);
         final Chunk endChunk = jsonbContext.unmarshall(jmsMessage.getText(), Chunk.class);
         assertThat("chunk ID", endChunk.getChunkId(), is(0L));
-        assertThat("chunk belongs to job", endChunk.getJobId(), is((long) jobInfoSnapshot.getJobId()));
+        assertThat("chunk belongs to job", endChunk.getJobId(), is(jobInfoSnapshot.getJobId()));
         assertThat("number of items in chunk", endChunk.getItems().size(), is(1));
         assertThat("chunk is termination chunk", endChunk.getItems().get(0).getType().get(0),
                 is(ChunkItem.Type.JOB_END));
